% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LIKE}
\alias{LIKE}
\title{Obtains likelihood}
\usage{
LIKE(Y, X, beta0, beta)
}
\arguments{
\item{Y}{Binary outcome vector.}

\item{X}{Matrix of covariates.}

\item{beta0}{Intercept parameter.}

\item{beta}{Vector of covariate parameters of length p.}
}
\value{
Likelihood
}
\description{
Calculates likelihood from observed outcome data and given covariate data/parameters.
}
